moved async clone tests into same class as other clone tests
